transfers: Do not Prune transfers in error. Keep them in a different list
When the limit of finished transfer is reach, transfers in error are
removed. We actually want to be able to retrieve the list of transfers
that failed. In order to do so instead of removing a transfer in error
move it to a second list which will only contain errors.